db clone rake tasks: fix db:*:clone task and add double non-production check

+++ lib/tasks/db.rake
@@ -25,35 +25,37 @@ namespace :db do
             "--user=#{config['username']} "
   end
 
-  ['staging', 'production'].each do |envname|
+  if ENV['RAILS_ENV'] != 'production'
+    ['staging', 'production'].each do |envname|
 
-    desc "Dump #{envname} DB into tmp/dump.#{envname}.sql.gz"
-    task "#{envname}:dump" do
-      raise 'Cannot run in a live env!' if ['staging', 'production'].include? Rails.env
-      config = get_db_config_for_env(envname)
-      mysqldump_cmd = "mysqldump " +
-              get_mysql_connection_options_from_config(config) +
-              "--skip-add-drop-table " +
-              "#{config['database']}"
-      system("cd #{Rails.root}; #{mysqldump_cmd} | gzip -c > tmp/dump.#{envname}.sql.gz") or raise("Failed to dump #{envname} DB")
+      desc "Dump #{envname} DB into tmp/dump.#{envname}.sql.gz"
+      task "#{envname}:dump" do
+        raise 'Cannot run in live env!' if Rails.env.production?
+        config = get_db_config_for_env(envname)
+        mysqldump_cmd = "mysqldump " +
+                get_mysql_connection_options_from_config(config) +
+                "--skip-add-drop-table " +
+                "#{config['database']}"
+        system("cd #{Rails.root}; #{mysqldump_cmd} | gzip -c > tmp/dump.#{envname}.sql.gz") or raise("Failed to dump #{envname} DB")
+      end
+
+      desc "Load tmp/dump.#{envname}.sql.gz into current environment DB"
+      task "#{envname}:load" => ['db:drop', 'db:create', :environment] do
+        raise 'Cannot run in live env!' if Rails.env.production?
+        config = get_db_config_for_env(Rails.env)
+
+        mysql_connection_options = get_mysql_connection_options_from_config(config)
+
+        mysqldump_cmd = "mysqldump " +
+                "--no-data --add-drop-table " +
+                mysql_connection_options
+        "#{config['database']}"
+        system("cd #{Rails.root}; zcat tmp/dump.#{envname}.sql.gz | mysql #{get_mysql_connection_options_from_config(config)} #{config['database']}") or raise("Load failed")
+      end
+
+      desc "Copy the contents of the #{envname} DB into the current environment DB"
+      task "#{envname}:clone" => ["#{envname}:dump", "#{envname}:load"]
     end
-
-    desc "Load tmp/dump.#{envname}.sql.gz into current environment DB"
-    task "#{envname}:load" => ['db:drop', 'db:create', :environment] do
-      raise 'Cannot run in a live env!' if ['staging', 'production'].include? Rails.env
-      config = get_db_config_for_env(Rails.env)
-
-      mysql_connection_options = get_mysql_connection_options_from_config(config)
-
-      mysqldump_cmd = "mysqldump " +
-              "--no-data --add-drop-table " +
-              mysql_connection_options
-              "#{config['database']}"
-      system("cd #{Rails.root}; zcat tmp/dump.#{envname}.sql.gz | mysql #{get_mysql_connection_options_from_config(config)} #{config['database']}") or raise("Load failed")
-    end
-
-    desc "Copy the contents of the #{envname} DB into the current environment DB"
-    task "#{envname}:clone" => ["#{envname}dump", "#{envname}:load"]
   end
 
 end
